+++ git_machete/generated_docs.py
@@ -628,9 +628,8 @@ long_docs: Dict[str, str] = {
               If `.git/info/description` or `.github/pull_request_template.md` file is present, its contents are used as PR description.
               Otherwise (or if `machete.github.forceDescriptionFromCommitMessage` is set), PR description is taken from message body of the first unique commit of the branch.
 
-              If the newly-created PR is stacked atop another PR, the actual PR description posted to GitHub will be prepended with the following header:
-
-              `# Based on PR #<number of PR for base branch>`
+              If the newly-created PR is stacked atop another PR, the actual PR description posted to GitHub will be prepended with a section
+              listing the entire related chain of PRs.
 
               <b>Options:</b>
 
@@ -659,9 +658,8 @@ long_docs: Dict[str, str] = {
  
               Sets the base of the current (or specified) branch's PR to upstream (parent) branch, as seen by git machete (see `git machete show up`).
 
-              If after changing the base, the PR ends up stacked atop another PR, the PR description at GitHub will be prepended with the following header:
-
-              `# Based on PR #<number of PR for base branch>`
+              If after changing the base the PR ends up stacked atop another PR, the PR description posted to GitHub will be prepended with a section
+              listing the entire related chain of PRs.
 
               This header will be updated or removed accordingly with the subsequent runs of `retarget-pr`.