+++ RELEASE_NOTES.md
@@ -1,8 +1,9 @@
 # Release notes
 
-## New in git-machete 3.22.1
+## New in git-machete 3.23.0
 
-- fixed: in the unlikely case of a cycle between GitHub PRs, `git machete github checkout-prs` aborts with an error rather than falling into an infinite loop
+- added: full chain of PRs (and not just a link to the base PR) is added to/updated in PR description by `git machete github create-pr`/`retarget-pr`/`restack-pr` (suggested by @mjgigli)
+- fixed: in the unlikely case of a cycle between GitHub PRs, `git machete github checkout-pr` aborts with an error rather than falling into an infinite loop
 - fixed: when checking out longer PR chains, `git machete github checkout-prs` prints out all checked out branches correctly
 
 ## New in git-machete 3.22.0